Update optimizely-tutorial.md
The existing instructions for SSO integration between Microsoft Entra and Optimizely were outdated. They referred to the SSO connection for the standalone Optimizely Experimentation platform.

Nowadays, the access to the multiple Optimizely products is centralized in Opti ID. The SSO configuration needs to be set up via the Optimizely Admin Center. The Optimizely knowledge base article is kept up-to-date, and so it's recommended to refer to that article, and not have outdated documentations that confuse customers.