(AzureCXP) fixes MicrosoftDocs/entra-docs#420678
Updated line number 19.

From: The following document guides you through on-demand provisioning with Microsoft Entra Cloud Sync for provisioning from Microsoft Entra ID to Active Directory.  If you are looking for information on provisioning from AD to Microsoft Entra ID, see [ On-demand provisioning - Microsoft Entra ID to Active Directory](how-to-on-demand-provision-entra-to-active-directory.md).

To: The following document guides you through on-demand provisioning with Microsoft Entra Cloud Sync for provisioning from Microsoft Entra ID to Active Directory.  If you are looking for information on provisioning from Microsoft Entra ID to AD, see [ On-demand provisioning - Microsoft Entra ID to Active Directory](how-to-on-demand-provision-entra-to-active-directory.md).