remove 'roles' attribute mapping in BIC Cloud Design provisioning guide
The BIC Cloud Design SCIM API is removing support for the 'roles' attribute due to limitations and customer confusion. Role assignment is instead handled via SSO based on group memberships.

This update removes the 'roles' attribute from the attribute mapping table in the provisioning documentation to reflect this change in the application's SCIM support.