Update v2-oauth2-client-creds-grant-flow.md
The proposed change is for this following paragraph - adding more clarity at the end of the paragraph:

"For a higher level of assurance, the Microsoft identity platform also allows the calling service to authenticate using a certificate or federated credential instead of a shared secret. Because the application's own credentials are being used, these credentials must be kept safe. Never publish that credential in your source code, embed it in web pages, or use it in a widely distributed native application."

The proposed change:
Adding "Authentication requests using the client credentials flow with web pages will not be allowed"

The reason for the change suggestion:
Customer is seeing ambiguity in this article, and they don't see any wording stating that "a client secret should not be passed in a web app for the flow".

Suggestion is to update the public documentation by explicitly stating that authentication requests using the client credentials flow with web pages will not be allowed.