Prevent users from reading other users information
While supporting a customer, I spent several hours (which is not desired) investigating how to prevent users from accessing other users' information in Microsoft Entra ID. This was a critical requirement due to the customer's internal policy to protect personally identifiable information (PII). As part of this effort, I identified that the allowedToReadOtherUsers property within the authorizationPolicy needed to be configured appropriately. I updated the documentation to include direct references to the relevant Microsoft Graph API resources to help others quickly locate and understand how to apply this setting.