Update manage-roles-portal.md
Request this update to clarify the restrictions that apply when the target user in an administrative unit has an Entra id role assigned. Previously, the doc mentioned that restriction applies only if targt user has admins roles assigned, but this is not accurate as the restriction also applies for other Entra Id roles, this was reviewed by EEEs on an IcM opened by CSS case.