Update concept-sspr-policy.md
We have tested few scenarios where the user that is created in the cloud with no on-prem footprint can always reset their password using their old password. No matter where they reset they can re-use the same password over and over even after expiration.

This is because Entra ID does not store user's password, therefore when a user tries to reset their password and use their immediate old password, they are still able to reset.

The table in this document indicates that it shouldn't be possible and customers are opening support cases on. But this only applies when the user is synced from on-prem and tries to do password writeback.

password change history	The last password can't be used again when the user changes a password.
Password reset history	The last password can be used again when the user resets a forgotten password.