Update github-enterprise-managed-user-ghe-com-tutorial.md
Updating these docs to direct people to the GitHub documentation for configuring their IDP within GitHub for an EMU enterprise. 

It's worth noting that GitHub Support cannot update these configurations from our side. This documentation is inaccurate in it's current form and it's imperative that we update this to reflect how customers should proceed.