Update groups-concept.md
We have removed the following item from the Known Issues section:

------------
If an administrator role is assigned to a role-assignable group instead of individual users, members of the group won't be able to access Rules, Organization, or Public Folders in the new Exchange admin center. The workaround is to assign the role directly to users instead of the group.
------------

After investigating the matter in collaboration with the EXO team, we confirmed that it is now possible to access Public Folders and Rules even when roles are assigned to groups.