Update how-to-prerequisites.md
I have added a note to inform users that it is not currently possible to check the configuration size, to determine, in larger environments, whether or not the configuration is close to or has exceeded the 4 MB size limit. 

I have included this after having a case open with Microsoft support for several months, trying to determine how to achieve this, with the official response coming back that its not currently possible.