Update how-to-unlock-users-for-mandatory-multifactor-authentication.md
fix the azure cli postponement script. 

The Token returns a SecureString but its usage was expecting a string so added a function to make that conversion from secure string to string

This script is used in the event of a tenant lockout due to mfa enforcement. 

Phase 1 of mfa enforcement is not applicable to azure cli so this script can be used to pospone phase 1 to sep 30, 2025 to regain access in the event of a lockout.