Clarify SMS support after legacy policy deprecation
Added clarification that SMS and other methods will remain functional under the Authentication methods policy despite the deprecation of legacy policies. The reason for this is customers are coming up with questions on this and we have an external blog written that is confusing customers even further. https://www.vissensa.com/blog/microsoft-mfa-microsoft-to-remove-sms-multi-factor-authentication/