(AzureCXP) fixes MicrosoftDocs/entra-docs-pr #508324
Changed line 49
From
In addition to the PRT, Microsoft Entra ID issues a Cloud TGT for the realm `KERBEROS.MICROSOFTONLINE.COM`. This is a partial TGT to access on-premises resources. In this model, Microsoft Entra ID acts as the KDC to facilitate seamless authentication.
To
In addition to the PRT, Microsoft Entra ID issues a Cloud TGT for the realm KERBEROS.MICROSOFTONLINE.COM, which is used for authentication to cloud resources. A separate partial TGT is issued for accessing on-premises resources, as described later in this document. In this model, Microsoft Entra ID acts as the KDC to facilitate seamless authentication.